[i2c] Increase maximum `FifoDepth` from 127 to 4095
This is a prerequisite for increasing the depth of some of the FIFOs
beyond 127.

This implies that the four FIFO levels can no longer be exposed to SW in
the single 32-bit `FIFO_STATUS` register.  Instead, each of the two
host-mode FIFO levels is now a 12-bit field in the `HOST_FIFO_STATUS`
register, and each of the two target-mode FIFO levels is now a 12-bit
field in the `TARGET_FIFO_STATUS` register.

+++ sw/device/lib/dif/dif_i2c.h
@@ -551,10 +551,10 @@ dif_result_t dif_i2c_override_sample_pins(const dif_i2c_t *i2c,
  */
 OT_WARN_UNUSED_RESULT
 dif_result_t dif_i2c_get_fifo_levels(const dif_i2c_t *i2c,
-                                     uint8_t *fmt_fifo_level,
-                                     uint8_t *rx_fifo_level,
-                                     uint8_t *tx_fifo_level,
-                                     uint8_t *acq_fifo_level);
+                                     uint16_t *fmt_fifo_level,
+                                     uint16_t *rx_fifo_level,
+                                     uint16_t *tx_fifo_level,
+                                     uint16_t *acq_fifo_level);
 
 /**
  * Pops an entry (a byte) off of the RX FIFO. Passing in `NULL` to the out-param
